Can install Maya on Windows

+++ Editor/Scripts/DCCIntegration/MayaIntegrator.cs
@@ -7,7 +7,7 @@ using UnityEngine;
 namespace UnityEditor.MeshSync {
 
 internal class MayaIntegrator : BaseDCCIntegrator {
-
+    
     internal MayaIntegrator(DCCToolInfo dccToolInfo) : base(dccToolInfo) { }
 
 //----------------------------------------------------------------------------------------------------------------------
@@ -22,12 +22,12 @@ internal class MayaIntegrator : BaseDCCIntegrator {
     {
 
         string tempPath = FileUtil.GetUniqueTempPathInProject();
+        
         Directory.CreateDirectory(tempPath);
         ZipUtility.UncompressFromZip(localPluginPath, null, tempPath);
 
         string srcRoot = Path.Combine(tempPath, Path.GetFileNameWithoutExtension(localPluginPath));
         if (!Directory.Exists(srcRoot)) {
-            Debug.LogError("[MeshSync] Failed to install DCC Plugin for Maya");
             return null;
         }
 
@@ -38,43 +38,67 @@ internal class MayaIntegrator : BaseDCCIntegrator {
 
         int exitCode = 0;
 
+        string copySrcFolder  = srcRoot;
+        string copyDestFolder = configFolder;
+        string argFormat = null;
+        string loadPluginCmd = null;
+            
+            
         switch (Application.platform) {
             case RuntimePlatform.WindowsEditor: {
-                // If MAYA_APP_DIR environment variable is setup, copy the modules directory there.
-                //     If not, go to %USERPROFILE%\Documents\maya in Windows Explorer, and copy the modules directory there.
-                //string commandString = "-command \"{0}\"";
+                //C:\Users\Unity\Documents\maya\modules
+                const string FOLDER_PREFIX = "modules";
+                copySrcFolder  = Path.Combine(srcRoot, FOLDER_PREFIX);
+                copyDestFolder = Path.Combine(configFolder, FOLDER_PREFIX);                
+                
+                argFormat = "-command \"{0}\"";
+
+                //Maya script only supports '/' as PathSeparator
+                //Example: loadPlugin """C:/Users/Unity/Documents/maya/modules/UnityMeshSync/2019/plug-ins/MeshSyncClientMaya.mll""";
+                string mayaPluginPath = Path.Combine(copyDestFolder, "UnityMeshSync", dccToolInfo.DCCToolVersion, 
+                    @"plug-ins\MeshSyncClientMaya.mll").Replace('\\','/');
+                loadPluginCmd = "loadPlugin \"\"\"" + mayaPluginPath + "\"\"\";";
                 break;
             }
             case RuntimePlatform.OSXEditor: {
                 
-                //Copy files
-                const string MOD_FILE = "UnityMeshSync.mod";
-                string SCRIPT_FOLDER = Path.Combine("UnityMeshSync",dccToolInfo.DCCToolVersion);
-                File.Copy(Path.Combine(srcRoot,MOD_FILE), Path.Combine(configFolder,MOD_FILE), true);
-                FileUtility.CopyRecursive(Path.Combine(srcRoot, SCRIPT_FOLDER), 
-                                          Path.Combine(configFolder, SCRIPT_FOLDER),
-                                          true);
-                
                 //Setup Auto Load
-                string commandString = @"-command '{0}'";
+                argFormat = @"-command '{0}'";
                 //Example: "/Users/Shared/Autodesk/Modules/maya/UnityMeshSync/2020/plug-ins/MeshSyncClientMaya.bundle";
-                string loadPlugin = "loadPlugin \"" + configFolder + "/UnityMeshSync/"+ dccToolInfo.DCCToolVersion 
-                                    + "/plug-ins/MeshSyncClientMaya.bundle\";";
-                
-                string argument = string.Format(commandString, loadPlugin+FINALIZE_SETUP);
-                exitCode = SetupAutoLoadPlugin(dccToolInfo.AppPath, argument);
+                loadPluginCmd = "loadPlugin \"" + configFolder + "/UnityMeshSync/"+ dccToolInfo.DCCToolVersion 
+                                       + "/plug-ins/MeshSyncClientMaya.bundle\";";
                 
                 break;
             }
             case RuntimePlatform.LinuxEditor: {
                 // Copy the modules directory to ~/maya/<maya_version)            
-                break;
+                throw new NotImplementedException();
             }
             default: {
                 throw new NotImplementedException();
             }
         }
-        
+
+        //Copy files
+        const string MOD_FILE = "UnityMeshSync.mod";
+        string scriptFolder = Path.Combine("UnityMeshSync",dccToolInfo.DCCToolVersion);
+        string srcModFile = Path.Combine(copySrcFolder, MOD_FILE);
+        if (!File.Exists(srcModFile)) {
+            return null;
+        }
+        try {
+            Directory.CreateDirectory(copyDestFolder);
+            File.Copy(srcModFile, Path.Combine(copyDestFolder, MOD_FILE), true);
+            FileUtility.CopyRecursive(Path.Combine(copySrcFolder, scriptFolder),
+                Path.Combine(copyDestFolder, scriptFolder),
+                true);
+        } catch {
+            return null;
+        }
+
+        //Auto Load
+        string arg = string.Format(argFormat, loadPluginCmd+FINALIZE_SETUP);
+        exitCode = SetupAutoLoadPlugin(dccToolInfo.AppPath, arg);
 
         //Cleanup
         FileUtility.DeleteFilesAndFolders(tempPath);
@@ -124,7 +148,7 @@ internal class MayaIntegrator : BaseDCCIntegrator {
     // [SecurityPermission(SecurityAction.LinkDemand, Flags = SecurityPermissionFlag.UnmanagedCode)]
     int SetupAutoLoadPlugin(string mayaPath, string startArgument) {
         int exitCode = 0;
-
+        
         try {
             if (!System.IO.File.Exists(mayaPath)) {
                 Debug.LogError("[MeshSync] No maya installation found at " + mayaPath);
